Update contract Metadata entrypoint and minor changes (#542)
## 1. Overview

This PR aims to add an entrypoint on CL Vault contract to update the
Metadata such as Name and Thesis. Included all msgs dependencies and
unit tests.

Also move the BuildTickCache entrypoint to the correct place, under
Admin execute messages and add a test tube test case for it.

## 2. Implementation details

<!-- Describe the implementation (highlights only) as well as design
rationale. -->

## 3. How to test/use

<!-- How can people test/use this? -->

## 4. Checklist

<!-- Checklist for PR author(s). -->

- [ ] Does the Readme need to be updated?

## 5. Limitations (optional)

<!-- Describe any limitation of the capabilities listed in the Overview
section. -->

## 6. Future Work (optional)

<!-- Describe follow up work, if any. -->