feat: rewrite deposit math, autocompounding, AnyDeposit(#586)
## 1. Overview

- Change ExactDeposit to no longer add to position, but just deposit on
ratio and refund the rest
- Add AnyDeposit, that swaps internally
- Add Autocompounding

## 2. Implementation details
### Exact Deposit
exact deposit no longer adds funds to the vaults current position, but
instead adds the funds to the vaults free balance. In doing so, the
vaults are more able to handle the dynamic Osmosis uptimes. This change
should also result in smaller refund amounts. Using the autocompounding
endpoints, these funds can later be added to the vaults position

### Any Deposit
This PR adds an endpoint for users to deposit any amount of token0 or
token1 into the vault. The vault then swaps these tokens in the needed
ratio. The swapped tokens then become part of the vaults free balance.
Using the autocompounding endpoints, these funds can later be added to
the vaults position

### Autocompounding
This PR adds multiple endpoints to the CL for autocompounding. The three
endpoints are:
1. CollectRewards, collects the accrued rewards from the position and
saves the amount
2. AutoCompoundRewards, this swaps the rewards to the current ratio of
the vault, preparing them for redepositing
3. Redeposit, this adds all funds to the position without doing any
further swapping
The implementations of exact depsoit and any deposit depend on redeposit
being present for adding funds to the vault


## 3. How to test/use

Each different feature can be tested fully in test tube. Autocompounding
can be fully tested separate from the deposit methods. The deposit
methods should include a call to redeposit for good measure.
For good measure, the rollout of this PR should be done in steps to make
sure all off chain infrastructure is properly adapted

## 4. Checklist

<!-- Checklist for PR author(s). -->

- [ ] Does the Readme need to be updated?
Yes

## 5. Limitations (optional)

<!-- Describe any limitation of the capabilities listed in the Overview
section. -->
- The current AnyDeposit does not allow for swap routes to be added to
the internal swap.
- Both deposit methods do not fully add a concept of uptimes

## 6. Future Work (optional)

<!-- Describe follow up work, if any. -->
A separate AnyDeposit method without a recipient field should added for
Authz integrations such as Yieldmos

---------

Co-authored-by: magiodev <31893902+magiodev@users.noreply.github.com>
Co-authored-by: Laurens <32776056+LaurensKubat@users.noreply.github.com>
Co-authored-by: Ajaz Ahmed Ansari <53616488+ajansari95@users.noreply.github.com>
Co-authored-by: Nikita (rizz) <nj18@nyu.edu>